`pub` `impl Arbitrary`
# Should the property test be `arb` feature-gated?
I inclinde to *no* because the tests are not feature-specific and it's a good thing `cargo t` fails until like `--features` is added.
# `#[cfg(any(test, feature = "arb"))]`
I considered this approach but having some experience with it shows that syncing versions between the feature and dev.dependencies is a nuisance. Being more naive it doesn't add much (only working `cargo t` but it works smoothly adding the feature).